+++ R/ggplot_rsi.R
@@ -71,7 +71,7 @@
 #'   select(hospital_id, amox, cipr) %>%
 #'   group_by(hospital_id) %>%
 #'   ggplot_rsi() +
-#'   facet_grid("hospital_id") +
+#'   facet_wrap("hospital_id", nrow = 1) +
 #'   labs(title = "AMR of Amoxicillin And Ciprofloxacine Per Hospital")
 ggplot_rsi <- function(data,
                        position = "stack",